requirements.txt
flask
keras-facenet
pillow
numpy
opencv-python
matplotlib
scipy
tensorflow